NEW FREEEEEE LESSSONNNNNNNNNNNNNNNNN!!!
Reverse engineering course that was released for free today!!! Happy to finally release this one. 

## New Notes 

- New theme 

We are working on a new theme if you could not tell. I started to prefer this one as it felt more professional and specific to the marketing- it also allowed us to make information display more formal. 

- New Content 

This article stems out from the reverse engineering series we released to start the first lesson series. This is not the full end of all RE content, but we have some more cool stuff coming- and this teases at a lot of new topics, and areas that can incorporate RE.